move types for camera, add subcomponents for model viewer snippet
