+++ packages/space-opera/src/components/camera_settings/camera_settings_test.ts
@@ -20,10 +20,10 @@ import './camera_settings.js';
 
 import {dispatchCurrentCameraState} from '../../redux/space_opera_base.js';
 import {reduxStore} from '../../redux/space_opera_base.js';
-import {Vector3D} from '../../redux/state_types.js';
 
 import {CameraSettings, CameraTargetInput} from './camera_settings.js';
 import {dispatchAutoRotate, dispatchCameraTarget, dispatchInitialOrbit} from './reducer.js';
+import {Vector3D} from './types.js';
 
 describe('camera constraints test', () => {
   let cameraSettings: CameraSettings;
