+++ packages/space-opera/src/components/model_viewer_snippet/components/open_button.ts
@@ -16,14 +16,14 @@
  */
 
 import '@material/mwc-button';
-import '../file_modal/file_modal.js';
+import '../../file_modal/file_modal.js';
 
 import {createSafeObjectUrlFromArrayBuffer} from '@google/model-viewer-editing-adapter/lib/util/create_object_url.js'
 import {customElement, html, LitElement, query} from 'lit-element';
 
-import {dispatchConfig, dispatchGltfUrl, extractStagingConfig, reduxStore} from '../../redux/space_opera_base.js';
-import {FileModalElement} from '../file_modal/file_modal.js';
-import {dispatchSetHotspots} from '../hotspot_panel/reducer.js';
+import {dispatchConfig, dispatchGltfUrl, extractStagingConfig, reduxStore} from '../../../redux/space_opera_base.js';
+import {FileModalElement} from '../../file_modal/file_modal.js';
+import {dispatchSetHotspots} from '../../hotspot_panel/reducer.js';
 
 /**
  * A button to open file resources.
