+++ packages/space-opera/src/components/model_viewer_snippet/model_viewer_snippet.ts
@@ -26,7 +26,6 @@ import {isObjectUrl} from '@google/model-viewer-editing-adapter/lib/util/create_
 import {css, customElement, html, internalProperty, LitElement, query} from 'lit-element';
 
 import {applyCameraEdits, Camera, INITIAL_CAMERA} from '../../redux/camera_state.js';
-import {parseHotspotsFromSnippet} from '../../redux/parse_hotspot_config.js';
 import {dispatchConfig, dispatchGltfUrl, State} from '../../redux/space_opera_base.js';
 import {ConnectedLitElement} from '../connected_lit_element/connected_lit_element.js';
 import {ExportZipButton} from '../download_button/download_button.js';
@@ -36,6 +35,8 @@ import {SnippetViewer} from '../shared/snippet_viewer/snippet_viewer.js';
 import {styles as hotspotStyles} from '../utils/hotspot/hotspot.css.js';
 import {renderHotspots} from '../utils/hotspot/render_hotspots.js';
 import {renderModelViewer} from '../utils/render_model_viewer.js';
+
+import {parseHotspotsFromSnippet} from './parse_hotspot_config.js';
 
 /**
  *
