+++ packages/space-opera/src/components/model_viewer_snippet/parse_hotspot_config_test.ts
